requirements.txt
requests>=2.32
python-dotenv>=1.0
pyodbc>=5.1
